requirements.txt
yfinance
pandas
numpy
google-cloud-storage
google-cloud-bigquery
pyarrow
functions-framework